server/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.6.0
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1

gunicorn==23.0.0
alembic==1.14.1

psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==5.0.0
cloudinary==1.44.1
flasgger==0.9.7.1